APIs/requirements.txt
python-doctr[torch]
fuzzywuzzy[speedup]
sentence-transformers
ultralytics
python-dotenv

flask
flask-cors
gunicorn==21.2.0

appwrite
